# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fornecedores API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/telefones": {
            "get": {
                "tags": [
                    "Telefones"
                ],
                "summary": "Listar todos os telefones",
                "operationId": "b61db6b665859d36b3105e4a284c8092",
                "responses": {
                    "200": {
                        "description": "Lista de telefones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Telefone"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Telefones"
                ],
                "summary": "Criar um novo telefone",
                "operationId": "4a149d88ef14c9bb2bbf5b5935a6071b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Telefone"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Telefone criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Telefone"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/telefones/{id}": {
            "get": {
                "tags": [
                    "Telefones"
                ],
                "summary": "Mostrar um telefone específico",
                "operationId": "3138fe9e2884ccd0f6ff04a9c96dc204",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do telefone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Telefone encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Telefone"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Telefone não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Telefones"
                ],
                "summary": "Atualizar um telefone",
                "operationId": "e073316d97717d4e4ce04d05abdfd2d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do telefone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Telefone"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Telefone atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Telefone"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Telefone não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Telefones"
                ],
                "summary": "Deletar um telefone",
                "operationId": "1856290a777cc09384b3b2fbc2c3b6ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do telefone",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Telefone deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Telefone deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Telefone não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Telefone": {
                "required": [
                    "telefone"
                ],
                "properties": {
                    "id": {
                        "description": "ID do telefone",
                        "type": "integer"
                    },
                    "telefone": {
                        "description": "Número de telefone",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Data de exclusão",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}